# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Construction Coordination Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/engineering/changes": {
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "Changes",
        "operationId": "changes_api_projects__project_id__engineering_changes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "revision_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Change"
                  },
                  "title": "Response Changes Api Projects  Project Id  Engineering Changes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/engineering/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_projects__project_id__engineering_evidence__evidence_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/engineering/findings": {
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "Findings",
        "operationId": "findings_api_projects__project_id__engineering_findings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Finding"
                  },
                  "title": "Response Findings Api Projects  Project Id  Engineering Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "engineering"
        ],
        "summary": "Propose",
        "operationId": "propose_api_projects__project_id__engineering_findings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Finding"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/engineering/findings/{finding_id}": {
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "Finding",
        "operationId": "finding_api_projects__project_id__engineering_findings__finding_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Finding"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/engineering/findings/{finding_id}/decisions": {
      "post": {
        "tags": [
          "engineering"
        ],
        "summary": "Decide",
        "operationId": "decide_api_projects__project_id__engineering_findings__finding_id__decisions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Finding"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/engineering/findings/{finding_id}/coordination": {
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "Coordination",
        "operationId": "coordination_api_projects__project_id__engineering_findings__finding_id__coordination_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Coordination"
                  },
                  "title": "Response Coordination Api Projects  Project Id  Engineering Findings  Finding Id  Coordination Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/engineering/findings/{finding_id}/rechecks": {
      "get": {
        "tags": [
          "engineering"
        ],
        "summary": "Rechecks",
        "operationId": "rechecks_api_projects__project_id__engineering_findings__finding_id__rechecks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReCheck"
                  },
                  "title": "Response Rechecks Api Projects  Project Id  Engineering Findings  Finding Id  Rechecks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "engineering"
        ],
        "summary": "Request Recheck",
        "operationId": "request_recheck_api_projects__project_id__engineering_findings__finding_id__rechecks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReCheckRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReCheck"
                  },
                  "title": "Response Request Recheck Api Projects  Project Id  Engineering Findings  Finding Id  Rechecks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Change": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "from_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Revision Id"
          },
          "to_revision_id": {
            "type": "string",
            "title": "To Revision Id"
          },
          "subject": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DrawingTarget"
              },
              {
                "$ref": "#/components/schemas/CadTarget"
              },
              {
                "$ref": "#/components/schemas/BimTarget"
              },
              {
                "$ref": "#/components/schemas/DocumentTarget"
              }
            ],
            "title": "Subject",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "bim": "#/components/schemas/BimTarget",
                "cad": "#/components/schemas/CadTarget",
                "document": "#/components/schemas/DocumentTarget",
                "drawing": "#/components/schemas/DrawingTarget"
              }
            }
          },
          "kind": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Kind"
          },
          "aspects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aspects",
            "default": []
          },
          "detector": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Detector"
          },
          "detector_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector Version"
          },
          "raw_artifact_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Artifact Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "source_id",
          "from_revision_id",
          "to_revision_id",
          "subject",
          "kind",
          "aspects",
          "detector",
          "detector_version",
          "raw_artifact_key",
          "created_at"
        ],
        "title": "Change"
      },
...
      "Coordination": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "decision": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "DISMISSED",
              "CLOSED",
              "EDITED"
            ],
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note",
            "default": ""
          },
          "recheck_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recheck Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "finding_id",
          "actor",
          "decision",
          "note",
          "recheck_id",
          "created_at"
        ],
        "title": "Coordination"
      },
...
      "Evidence": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_revision": {
            "type": "string",
            "title": "Source Revision"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "work_package_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Package Id"
          },
          "element_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Element Ids",
            "default": []
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "fact": {
            "type": "string",
            "title": "Fact"
          },
          "quality": {
            "type": "string",
            "enum": [
              "structured",
              "extracted",
              "inferred"
            ],
            "title": "Quality",
            "default": "structured"
          },
          "source_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision Id"
          },
          "viewer_target": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DrawingTarget"
                  },
                  {
                    "$ref": "#/components/schemas/CadTarget"
                  },
                  {
                    "$ref": "#/components/schemas/BimTarget"
                  },
                  {
                    "$ref": "#/components/schemas/DocumentTarget"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "bim": "#/components/schemas/BimTarget",
                    "cad": "#/components/schemas/CadTarget",
                    "document": "#/components/schemas/DocumentTarget",
                    "drawing": "#/components/schemas/DrawingTarget"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "snapshot_id",
          "provider",
          "source_id",
          "source_revision",
          "observed_at",
          "work_package_id",
          "element_ids",
          "page",
          "location",
          "fact",
          "quality",
          "source_revision_id",
          "viewer_target"
        ],
        "title": "Evidence"
      },
...
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "work_package_id": {
            "type": "string",
            "title": "Work Package Id"
          },
          "conclusion": {
            "type": "string",
            "title": "Conclusion"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "reasoning_summary": {
            "type": "string",
            "title": "Reasoning Summary"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "impact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Impact"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "what_changed": {
            "type": "string",
            "title": "What Changed",
            "default": ""
          },
          "why_it_matters": {
            "type": "string",
            "title": "Why It Matters",
            "default": ""
          },
          "change_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Change Ids",
            "default": []
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/FindingDependency"
            },
            "type": "array",
            "title": "Dependencies",
            "default": []
          },
          "suggested_discipline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Discipline"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Action"
          },
          "state": {
            "type": "string",
            "enum": [
              "PROPOSED",
              "CONFIRMED",
              "DISMISSED",
              "CLOSED"
            ],
            "title": "State",
            "default": "PROPOSED"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "snapshot_id",
          "work_package_id",
          "conclusion",
          "evidence_ids",
          "reasoning_summary",
          "confidence",
          "limitations",
          "created_at",
          "project_id",
          "impact",
          "title",
          "what_changed",
          "why_it_matters",
          "change_ids",
          "dependencies",
          "suggested_discipline",
          "suggested_action",
          "state",
          "updated_at"
        ],
        "title": "Finding"
      },
      "FindingDecision": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "DISMISSED",
              "CLOSED",
              "EDITED"
            ],
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note",
            "default": ""
          },
          "recheck_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recheck Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "FindingDecision"
      },
...
      "FindingDraft": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Title"
          },
          "what_changed": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "What Changed"
          },
          "why_it_matters": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Why It Matters"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Evidence Ids"
          },
          "change_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Change Ids",
            "default": []
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/FindingDependency"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Dependencies"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "suggested_discipline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Discipline"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Action"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "what_changed",
          "why_it_matters",
          "evidence_ids",
          "dependencies",
          "impact"
        ],
        "title": "FindingDraft"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReCheckRequest": {
        "properties": {
          "operation_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Operation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id"
        ],
        "title": "ReCheckRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}